Create 00-welcome.md
First attempt to address [issue #231](https://github.com/swcarpentry/instructor-training/issues/231) on adding a welcome episode. 

These are the first three multiple questions that come to mind. A variation of the 2nd is part of the post-workshop questionnaire. I'm open to adding more or rewording these or any other suggestions. 

Also, I call this lesson 00, but it could be 01, especially if the current episode one is moved to later.